backend/requirements.txt
fastapi
python-dotenv
browser-use